t/data/parsing/temma-node-2.dat

#data
<t:text t:parse></t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>
|       temma parse=""

#data
<t:text t:parse>   </t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>
|       temma parse=""
|       "   "

#data
<t:text t:parse>abc</t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>
|       temma parse=""
|       <temma field>
|         "abc"

#data
<t:text t:parse><p>abc</t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>
|       temma parse=""
|       <temma field>
|         <p>
|           "abc"

#data
<t:text t:parse>abc<t:field></t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>
|       temma parse=""
|       <temma field>
|         "abc"
|       <temma field>

#data
<t:text t:parse>ab<p>c<t:field></t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>
|       temma parse=""
|       <temma field>
|         "ab"
|         <p>
|           "c"
|       <temma field>

#data
<t:text t:parse><t:field></t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>
|       temma parse=""
|       <temma field>

#data
<t:text t:parse><t:field>xyz<t:text></t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>
|       temma parse=""
|       <temma field>
|         "xyz"
|         <temma text>

#data
<t:barehtml t:parse></t:barehtml>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma barehtml>
|       temma parse=""

#data
<t:barehtml t:parse>   </t:barehtml>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma barehtml>
|       temma parse=""
|       "   "

#data
<t:barehtml t:parse>abc</t:barehtml>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma barehtml>
|       temma parse=""
|       <temma field>
|         "abc"

#data
<t:barehtml t:parse><p>abc</t:barehtml>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma barehtml>
|       temma parse=""
|       <temma field>
|         <p>
|           "abc"

#data
<t:barehtml t:parse>abc<t:field></t:barehtml>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma barehtml>
|       temma parse=""
|       <temma field>
|         "abc"
|       <temma field>

#data
<t:barehtml t:parse>ab<p>c<t:field></t:barehtml>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma barehtml>
|       temma parse=""
|       <temma field>
|         "ab"
|         <p>
|           "c"
|       <temma field>

#data
<t:barehtml t:parse><t:field></t:barehtml>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma barehtml>
|       temma parse=""
|       <temma field>

#data
<t:barehtml t:parse><t:field>xyz<t:barehtml></t:barehtml>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma barehtml>
|       temma parse=""
|       <temma field>
|         "xyz"
|         <temma barehtml>
